Separate out featurespec tests, and runs them under dotty build as well (skipped some tests).
