+++ build.sbt
@@ -16,15 +16,19 @@ lazy val scalacticTestDotty = ScalatestBuild.scalacticTestDotty
 lazy val scalatest = ScalatestBuild.scalatest
 lazy val scalatestTest = ScalatestBuild.scalatestTest
 lazy val scalatestDiagramsTest = ScalatestBuild.scalatestDiagramsTest
+lazy val scalatestFeatureSpecTest = ScalatestBuild.scalatestFeatureSpecTest
 lazy val scalatestJS = ScalatestBuild.scalatestJS
 lazy val scalatestNative = ScalatestBuild.scalatestNative
 lazy val scalatestDotty = ScalatestBuild.scalatestDotty
 lazy val scalatestTestJS = ScalatestBuild.scalatestTestJS
 lazy val scalatestDiagramsTestJS = ScalatestBuild.scalatestDiagramsTestJS
+lazy val scalatestFeatureSpecTestJS = ScalatestBuild.scalatestFeatureSpecTestJS
 lazy val scalatestTestNative = ScalatestBuild.scalatestTestNative
 lazy val scalatestDiagramsTestNative = ScalatestBuild.scalatestDiagramsTestNative
+lazy val scalatestFeatureSpecTestNative = ScalatestBuild.scalatestFeatureSpecTestNative
 lazy val scalatestTestDotty = ScalatestBuild.scalatestTestDotty
 lazy val scalatestDiagramsTestDotty = ScalatestBuild.scalatestDiagramsTestDotty
+lazy val scalatestFeatureSpecTestDotty = ScalatestBuild.scalatestFeatureSpecTestDotty
 lazy val scalatestApp = ScalatestBuild.scalatestApp
 lazy val scalatestAppJS = ScalatestBuild.scalatestAppJS
 lazy val scalatestAppNative = ScalatestBuild.scalatestAppNative
