+++ jvm/featurespec-test/src/test/scala/org/scalatest/featurespec/AsyncFeatureSpecSpec.scala
@@ -13,27 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.featurespec
 
-import SharedHelpers.{EventRecordingReporter, thisLineNumber}
-import org.scalatest.events.{InfoProvided, MarkupProvided}
+import org.scalatest.SharedHelpers.{EventRecordingReporter, thisLineNumber}
 import scala.concurrent.{Promise, ExecutionContext, Future}
 import org.scalatest.concurrent.SleepHelper
+import org.scalatest.events.{InfoProvided, MarkupProvided}
 import org.scalatest.exceptions.{DuplicateTestNameException, NotAllowedException}
 import org.scalactic.Prettifier
 
 import scala.util.Success
 import org.scalatest.featurespec.{ AsyncFeatureSpec, AsyncFeatureSpecLike }
+import org.scalatest.funspec
+import org.scalatest.ParallelTestExecution
+import org.scalatest.Args
+import org.scalatest.Assertion
+import org.scalatest.Succeeded
+import org.scalatest.FailureMessages
+import org.scalatest.UnquotedString
 
-class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
+class FixtureAsyncFeatureSpecSpec extends funspec.AnyFunSpec {
 
   private val prettifier = Prettifier.default
 
-  describe("AsyncFeatureSpecLike") {
+  describe("AsyncFeatureSpec") {
 
+    // SKIP-DOTTY-START
+    // ParallelTestExecution not working yet.
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSpec extends AsyncFeatureSpecLike with ParallelTestExecution {
+      class ExampleSpec extends AsyncFeatureSpec with ParallelTestExecution {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -60,13 +69,13 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
 
         Scenario("test 4") {
           Future {
-            cancel
+            cancel()
           }
         }
 
         ignore("test 5") {
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -95,7 +104,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSpec extends AsyncFeatureSpecLike with ParallelTestExecution {
+      class ExampleSpec extends AsyncFeatureSpec with ParallelTestExecution {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -115,11 +124,11 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
         }
 
         Scenario("test 4") {
-          cancel
+          cancel()
         }
 
         ignore("test 5") {
-          cancel
+          cancel()
         }
 
         override def newInstance = new ExampleSpec
@@ -144,12 +153,13 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       assert(rep.testIgnoredEventsReceived.length == 1)
       assert(rep.testIgnoredEventsReceived(0).testName == "Scenario: test 5")
     }
+    // SKIP-DOTTY-END
 
     it("should run tests that return Future in serial by default") {
 
       @volatile var count = 0
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -196,13 +206,13 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         Scenario("test 1") {
-          SleepHelper.sleep(30)
+          SleepHelper.sleep(3000)
           assert(count == 0)
           count = 1
           succeed
@@ -210,7 +220,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
 
         Scenario("test 2") {
           assert(count == 1)
-          SleepHelper.sleep(50)
+          SleepHelper.sleep(5000)
           count = 2
           succeed
         }
@@ -234,14 +244,14 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     // SKIP-SCALATESTJS,NATIVE-START
-    it("should run tests and its future in same main thread when use SerialExecutionContext") {
+    it("should run tests and its future in same main thread when using SerialExecutionContext") {
 
       var mainThread = Thread.currentThread
       var test1Thread: Option[Thread] = None
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         Scenario("test 1") {
           Future {
@@ -256,7 +266,6 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
             succeed
           }
         }
-
       }
 
       val rep = new EventRecordingReporter
@@ -281,7 +290,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         Scenario("test 1") {
           val promise = Promise[Assertion]
@@ -316,7 +325,6 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
             s
           }
         }
-
       }
 
       val rep = new EventRecordingReporter
@@ -335,42 +343,9 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       assert(onCompleteThread.get == mainThread)
     }
 
-    it("should not hang if a Future completes with nothing on the queue") {
-      // Despite not having a notify registered as a callback on the Future
-      // passed to SerialExecutionContext.runNow, it is not possible to get
-      // a test to hang, because there's always one more transformation after
-      // the Future[Assertion] completes, and that's the one that transforms
-      // the Future[Assertion] to a Future[Outcome]. That will cause a
-      // job to be enqueued via SerialExecutionContext.execute, and that
-      // will do the final notify. After running that job, the Future passed
-      // to runNow will be complete, and runNow will return.
-      class ExampleSpec extends AsyncFeatureSpecLike {
-
-        Scenario("test 1") {
-          val promise = Promise[Assertion]
-          val timer = new java.util.Timer
-          timer.schedule(
-            new java.util.TimerTask {
-              def run(): Unit = {
-                promise.complete(Success(succeed))
-              }
-            },
-            1000
-          )
-          promise.future
-        }
-      }
-
-      val rep = new EventRecordingReporter
-      val suite = new ExampleSpec
-      val status = suite.run(None, Args(reporter = rep))
-      status.waitUntilCompleted()
-      assert(!rep.testSucceededEventsReceived.isEmpty)
-    }
-
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -398,7 +373,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -422,7 +397,6 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
             succeed
           }
         }
-
       }
 
       val rep = new EventRecordingReporter
@@ -444,7 +418,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends AsyncFeatureSpecLike {
+      class ExampleSpec extends AsyncFeatureSpec {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -462,7 +436,6 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
         Scenario("test 3") {
           succeed
         }
-
       }
 
       val rep = new EventRecordingReporter
@@ -483,7 +456,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
         info(
           "hi there"
         )
@@ -502,7 +475,10 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in feature body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
+
+        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
+//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         Feature("test feature") {
           info(
@@ -526,7 +502,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -558,7 +534,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an InfoProvided event for an info in Future returned by scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -592,7 +568,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
         note(
           "hi there"
         )
@@ -611,7 +587,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a NoteProvided event for a note in feature body") {
-      class MySuite extends AsyncFeatureSpecLike {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -621,15 +597,12 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
             "hi there"
           )
 
-          Scenario("test 1") {
-            succeed
-          }
+          Scenario("test 1") { succeed }
         }
       }
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-
       // SKIP-SCALATESTJS,NATIVE-START
       status.waitUntilCompleted()
       // SKIP-SCALATESTJS,NATIVE-END
@@ -640,44 +613,8 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       assert(noteList(0).message == "hi there")
     }
 
-    it("should report as failed test when non-fatal exception is thrown from scenario body") {
-
-      class ExampleSpec extends AsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
-
-        Scenario("test 1") {
-          succeed
-        }
-
-        Scenario("test 2") {
-          throw new IllegalStateException("oops!")
-        }
-
-        Scenario("test 3") {
-          succeed
-        }
-
-      }
-
-      val rep = new EventRecordingReporter
-      val suite = new ExampleSpec
-      val status = suite.run(None, Args(reporter = rep))
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "Scenario: test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "Scenario: test 3")
-      assert(rep.testSucceededEventsReceived.length == 2)
-      assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "Scenario: test 3")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "Scenario: test 2")
-    }
-
     it("should send a NoteProvided event for a note in scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -702,7 +639,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a NoteProvided event for a note in Future returned by scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -729,7 +666,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
         alert(
           "hi there"
         )
@@ -748,7 +685,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in feature body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -764,7 +701,6 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-
       // SKIP-SCALATESTJS,NATIVE-START
       status.waitUntilCompleted()
       // SKIP-SCALATESTJS,NATIVE-END
@@ -775,46 +711,8 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       assert(alertList(0).message == "hi there")
     }
 
-    it("should report as failed test when non-fatal exception is thrown from Future returned by scenario body") {
-
-      class ExampleSpec extends AsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
-
-        Scenario("test 1") {
-          succeed
-        }
-
-        Scenario("test 2") {
-          Future {
-            throw new IllegalStateException("oops!")
-          }
-        }
-
-        Scenario("test 3") {
-          succeed
-        }
-
-      }
-
-      val rep = new EventRecordingReporter
-      val suite = new ExampleSpec
-      val status = suite.run(None, Args(reporter = rep))
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "Scenario: test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "Scenario: test 3")
-      assert(rep.testSucceededEventsReceived.length == 2)
-      assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "Scenario: test 3")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "Scenario: test 2")
-    }
-
     it("should send an AlertProvided event for an alert in scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -839,7 +737,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in Future returned by scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -866,7 +764,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
         markup(
           "hi there"
         )
@@ -885,7 +783,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in feature body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -912,7 +810,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -944,7 +842,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by scenario body") {
-      class MySuite extends AsyncFeatureSpecLike  {
+      class MySuite extends AsyncFeatureSpec  {
 
         //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
 //SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
@@ -977,68 +875,6 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       assert(markupProvided.text == "hi there")
     }
 
-
-
-    // SKIP-SCALATESTJS,NATIVE-START
-    it("should propagate fatal exception when thrown from scenario body") {
-
-      class ExampleSpec extends AsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
-
-        Scenario("test 1") {
-          succeed
-        }
-
-        Scenario("test 2") {
-          throw new VirtualMachineError("oops!") {}
-        }
-
-        Scenario("test 3") {
-          succeed
-        }
-
-      }
-
-      val rep = new EventRecordingReporter
-      assertThrows[VirtualMachineError] {
-        val suite = new ExampleSpec
-        suite.run(None, Args(reporter = rep))
-      }
-    }
-
-    it("should propagate fatal exception when thrown from Future returned by scenario body") {
-
-      class ExampleSpec extends AsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
-
-        Scenario("test 1") {
-          succeed
-        }
-
-        Scenario("test 2") {
-          Future {
-            throw new VirtualMachineError("oops!") {}
-          }
-        }
-
-        Scenario("test 3") {
-          succeed
-        }
-
-      }
-
-      val rep = new EventRecordingReporter
-      assertThrows[VirtualMachineError] {
-        val suite = new ExampleSpec
-        suite.run(None, Args(reporter = rep))
-      }
-    }
-    // SKIP-SCALATESTJS,NATIVE-END
-
     it("should generate NotAllowedException wrapping a DuplicateTestNameException is thrown inside scope") {
       class TestSpec extends AsyncFeatureSpec {
         Feature("a feature") {
@@ -1049,7 +885,7 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       val e = intercept[NotAllowedException] {
           new TestSpec
         }
-      assert("AsyncFeatureSpecLikeSpec.scala" == e.failedCodeFileName.get)
+      assert("AsyncFeatureSpecSpec.scala" == e.failedCodeFileName.get)
       assert(e.failedCodeLineNumber.get == thisLineNumber - 7)
       assert(e.cause.isDefined)
       val causeThrowable = e.cause.get
@@ -1095,7 +931,5 @@ class AsyncFeatureSpecLikeSpec extends funspec.AnyFunSpec {
       assert(reporter.testStartingEventsReceived.length == 3)
       assert(reporter.testSucceededEventsReceived.length == 3)
     }
-
   }
-
 }
