+++ jvm/featurespec-test/src/test/scala/org/scalatest/featurespec/AsyncFeatureSpecSpec2.scala
@@ -13,20 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatest.featurespec
 
-import SharedHelpers.EventRecordingReporter
+import org.scalatest.SharedHelpers.EventRecordingReporter
 import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest.concurrent.SleepHelper
 import org.scalatest.events.{InfoProvided, MarkupProvided}
 
 import scala.util.Success
 import org.scalatest.featurespec.AsyncFeatureSpec
+import org.scalatest.funspec
+import org.scalatest.ParallelTestExecution
+import org.scalatest.Args
+import org.scalatest.Assertion
+import org.scalatest.Succeeded
 
-class AsyncFeatureSpecSpec2 extends funspec.AsyncFunSpec {
+class FixtureAsyncFeatureSpecSpec2 extends funspec.AsyncFunSpec {
 
   describe("AsyncFeatureSpec") {
 
+    // SKIP-DOTTY-START
+    // ParallelTestExecution not working yet.
     it("can be used for tests that return Future under parallel async test execution") {
 
       class ExampleSpec extends AsyncFeatureSpec with ParallelTestExecution {
@@ -53,13 +60,13 @@ class AsyncFeatureSpecSpec2 extends funspec.AsyncFunSpec {
 
         Scenario("test 4") {
           Future {
-            cancel
+            cancel()
           }
         }
 
         ignore("test 5") {
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -105,11 +112,11 @@ class AsyncFeatureSpecSpec2 extends funspec.AsyncFunSpec {
         }
 
         Scenario("test 4") {
-          cancel
+          cancel()
         }
 
         ignore("test 5") {
-          cancel
+          cancel()
         }
 
         override def newInstance = new ExampleSpec
@@ -134,6 +141,7 @@ class AsyncFeatureSpecSpec2 extends funspec.AsyncFunSpec {
         assert(repo.testIgnoredEventsReceived(0).testName == "Scenario: test 5")
       }
     }
+    // SKIP-DOTTY-END
 
     it("should run tests that return Future in serial by default") {
