+++ jvm/featurespec-test/src/test/scala/org/scalatest/featurespec/FixtureAsyncFeatureSpecLikeSpec2.scala
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.featurespec
 
-import scala.concurrent.{Promise, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Promise, Future}
 import org.scalatest._
 import SharedHelpers.EventRecordingReporter
 import org.scalatest.concurrent.SleepHelper
@@ -25,16 +25,15 @@ import scala.util.Success
 import org.scalatest
 import org.scalatest.featurespec
 
-class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
+class AsyncFeatureSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
   describe("AsyncFeatureSpecLike") {
 
+    // SKIP-DOTTY-START
+    // ParallelTestExecution not working yet.
     it("can be used for tests that return Future under parallel async test execution") {
 
       class ExampleSpec extends featurespec.FixtureAsyncFeatureSpecLike with ParallelTestExecution {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -62,13 +61,13 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
 
         Scenario("test 4") { fixture =>
           Future {
-            cancel
+            cancel()
           }
         }
 
         ignore("test 5") { fixture =>
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -78,28 +77,26 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val spec = new ExampleSpec
       val status = spec.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(rep.testStartingEventsReceived.length == 4)
-      assert(rep.testSucceededEventsReceived.length == 1)
-      assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "Scenario: test 2")
-      assert(rep.testPendingEventsReceived.length == 1)
-      assert(rep.testPendingEventsReceived(0).testName == "Scenario: test 3")
-      assert(rep.testCanceledEventsReceived.length == 1)
-      assert(rep.testCanceledEventsReceived(0).testName == "Scenario: test 4")
-      assert(rep.testIgnoredEventsReceived.length == 1)
-      assert(rep.testIgnoredEventsReceived(0).testName == "Scenario: test 5")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 4)
+        assert(repo.testSucceededEventsReceived.length == 1)
+        assert(repo.testSucceededEventsReceived(0).testName == "Scenario: test 1")
+        assert(repo.testFailedEventsReceived.length == 1)
+        assert(repo.testFailedEventsReceived(0).testName == "Scenario: test 2")
+        assert(repo.testPendingEventsReceived.length == 1)
+        assert(repo.testPendingEventsReceived(0).testName == "Scenario: test 3")
+        assert(repo.testCanceledEventsReceived.length == 1)
+        assert(repo.testCanceledEventsReceived(0).testName == "Scenario: test 4")
+        assert(repo.testIgnoredEventsReceived.length == 1)
+        assert(repo.testIgnoredEventsReceived(0).testName == "Scenario: test 5")
+      }
     }
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
       class ExampleSpec extends featurespec.FixtureAsyncFeatureSpecLike with ParallelTestExecution {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -120,11 +117,11 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
         }
 
         Scenario("test 4") { fixture =>
-          cancel
+          cancel()
         }
 
         ignore("test 5") { fixture =>
-          cancel
+          cancel()
         }
 
         override def newInstance = new ExampleSpec
@@ -133,30 +130,29 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val spec = new ExampleSpec
       val status = spec.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(rep.testStartingEventsReceived.length == 4)
-      assert(rep.testSucceededEventsReceived.length == 1)
-      assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testFailedEventsReceived.length == 1)
-      assert(rep.testFailedEventsReceived(0).testName == "Scenario: test 2")
-      assert(rep.testPendingEventsReceived.length == 1)
-      assert(rep.testPendingEventsReceived(0).testName == "Scenario: test 3")
-      assert(rep.testCanceledEventsReceived.length == 1)
-      assert(rep.testCanceledEventsReceived(0).testName == "Scenario: test 4")
-      assert(rep.testIgnoredEventsReceived.length == 1)
-      assert(rep.testIgnoredEventsReceived(0).testName == "Scenario: test 5")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 4)
+        assert(repo.testSucceededEventsReceived.length == 1)
+        assert(repo.testSucceededEventsReceived(0).testName == "Scenario: test 1")
+        assert(repo.testFailedEventsReceived.length == 1)
+        assert(repo.testFailedEventsReceived(0).testName == "Scenario: test 2")
+        assert(repo.testPendingEventsReceived.length == 1)
+        assert(repo.testPendingEventsReceived(0).testName == "Scenario: test 3")
+        assert(repo.testCanceledEventsReceived.length == 1)
+        assert(repo.testCanceledEventsReceived(0).testName == "Scenario: test 4")
+        assert(repo.testIgnoredEventsReceived.length == 1)
+        assert(repo.testIgnoredEventsReceived(0).testName == "Scenario: test 5")
+      }
     }
+    // SKIP-DOTTY-END
 
     it("should run tests that return Future in serial by default") {
 
       @volatile var count = 0
 
       class ExampleSpec extends featurespec.FixtureAsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -167,7 +163,7 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
             SleepHelper.sleep(30)
             assert(count == 0)
             count = 1
-            succeed
+            Succeeded
           }
         }
 
@@ -176,7 +172,7 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
             assert(count == 1)
             SleepHelper.sleep(50)
             count = 2
-            succeed
+            Succeeded
           }
         }
 
@@ -191,13 +187,12 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived.length == 3)
-
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 3)
+        assert(repo.testSucceededEventsReceived.length == 3)
+      }
     }
 
     it("should run tests that does not return Future in serial by default") {
@@ -205,9 +200,6 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       @volatile var count = 0
 
       class ExampleSpec extends featurespec.FixtureAsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -217,14 +209,14 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
           SleepHelper.sleep(30)
           assert(count == 0)
           count = 1
-          succeed
+          Succeeded
         }
 
         Scenario("test 2") { fixture =>
           assert(count == 1)
           SleepHelper.sleep(50)
           count = 2
-          succeed
+          Succeeded
         }
 
         Scenario("test 3") { fixture =>
@@ -236,13 +228,12 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived.length == 3)
-
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(repo.testStartingEventsReceived.length == 3)
+        assert(repo.testSucceededEventsReceived.length == 3)
+      }
     }
 
     // SKIP-SCALATESTJS,NATIVE-START
@@ -281,14 +272,17 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       status.whenCompleted { s =>
         onCompleteThread = Some(Thread.currentThread)
       }
-      status.waitUntilCompleted()
 
-      assert(test1Thread.isDefined)
-      assert(test1Thread.get == mainThread)
-      assert(test2Thread.isDefined)
-      assert(test2Thread.get == mainThread)
-      assert(onCompleteThread.isDefined)
-      assert(onCompleteThread.get == mainThread)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(test1Thread.isDefined)
+        assert(test1Thread.get == mainThread)
+        assert(test2Thread.isDefined)
+        assert(test2Thread.get == mainThread)
+        assert(onCompleteThread.isDefined)
+        assert(onCompleteThread.get == mainThread)
+      }
     }
 
     it("should run tests and its true async future in the same thread when use SerialExecutionContext") {
@@ -345,14 +339,17 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       status.whenCompleted { s =>
         onCompleteThread = Some(Thread.currentThread)
       }
-      status.waitUntilCompleted()
 
-      assert(test1Thread.isDefined)
-      assert(test1Thread.get == mainThread)
-      assert(test2Thread.isDefined)
-      assert(test2Thread.get == mainThread)
-      assert(onCompleteThread.isDefined)
-      assert(onCompleteThread.get == mainThread)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(test1Thread.isDefined)
+        assert(test1Thread.get == mainThread)
+        assert(test2Thread.isDefined)
+        assert(test2Thread.get == mainThread)
+        assert(onCompleteThread.isDefined)
+        assert(onCompleteThread.get == mainThread)
+      }
     }
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
@@ -382,17 +379,18 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      status.waitUntilCompleted()
-      assert(!rep.testSucceededEventsReceived.isEmpty)
+
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(!rep.testSucceededEventsReceived.isEmpty)
+      }
     }
     // SKIP-SCALATESTJS,NATIVE-END
 
     it("should run tests that returns Future and report their result in serial") {
 
       class ExampleSpec extends featurespec.FixtureAsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -423,26 +421,24 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "Scenario: test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "Scenario: test 3")
-      assert(rep.testSucceededEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "Scenario: test 2")
-      assert(rep.testSucceededEventsReceived(2).testName == "Scenario: test 3")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(rep.testStartingEventsReceived.length == 3)
+        assert(rep.testStartingEventsReceived(0).testName == "Scenario: test 1")
+        assert(rep.testStartingEventsReceived(1).testName == "Scenario: test 2")
+        assert(rep.testStartingEventsReceived(2).testName == "Scenario: test 3")
+        assert(rep.testSucceededEventsReceived.length == 3)
+        assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
+        assert(rep.testSucceededEventsReceived(1).testName == "Scenario: test 2")
+        assert(rep.testSucceededEventsReceived(2).testName == "Scenario: test 3")
+      }
     }
 
     it("should run tests that does not return Future and report their result in serial") {
 
       class ExampleSpec extends featurespec.FixtureAsyncFeatureSpecLike {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -467,18 +463,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val rep = new EventRecordingReporter
       val suite = new ExampleSpec
       val status = suite.run(None, Args(reporter = rep))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      assert(rep.testStartingEventsReceived.length == 3)
-      assert(rep.testStartingEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testStartingEventsReceived(1).testName == "Scenario: test 2")
-      assert(rep.testStartingEventsReceived(2).testName == "Scenario: test 3")
-      assert(rep.testSucceededEventsReceived.length == 3)
-      assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
-      assert(rep.testSucceededEventsReceived(1).testName == "Scenario: test 2")
-      assert(rep.testSucceededEventsReceived(2).testName == "Scenario: test 3")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(rep) }
+      promise.future.map { repo =>
+        assert(rep.testStartingEventsReceived.length == 3)
+        assert(rep.testStartingEventsReceived(0).testName == "Scenario: test 1")
+        assert(rep.testStartingEventsReceived(1).testName == "Scenario: test 2")
+        assert(rep.testStartingEventsReceived(2).testName == "Scenario: test 3")
+        assert(rep.testSucceededEventsReceived.length == 3)
+        assert(rep.testSucceededEventsReceived(0).testName == "Scenario: test 1")
+        assert(rep.testSucceededEventsReceived(1).testName == "Scenario: test 2")
+        assert(rep.testSucceededEventsReceived(2).testName == "Scenario: test 3")
+      }
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
@@ -495,21 +492,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
 
-      assert(infoList.size == 1)
-      assert(infoList(0).message == "hi there")
+        assert(infoList.size == 1)
+        assert(infoList(0).message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in feature body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -526,21 +521,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
 
-      assert(infoList.size == 1)
-      assert(infoList(0).message == "hi there")
+        assert(infoList.size == 1)
+        assert(infoList(0).message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -556,27 +549,25 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
-      assert(infoList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
+        assert(infoList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[InfoProvided])
-      val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
-      assert(infoProvided.message == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[InfoProvided])
+        val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
+        assert(infoProvided.message == "hi there")
+      }
     }
 
     it("should send an InfoProvided event for an info in Future returned by scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -594,27 +585,25 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val infoList = reporter.infoProvidedEventsReceived
-      assert(infoList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val infoList = reporter.infoProvidedEventsReceived
+        assert(infoList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[InfoProvided])
-      val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
-      assert(infoProvided.message == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[InfoProvided])
+        val infoProvided = recordedEvent.asInstanceOf[InfoProvided]
+        assert(infoProvided.message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -627,21 +616,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
 
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in feature body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -658,21 +645,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
 
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -688,20 +673,18 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send a NoteProvided event for a note in Future returned by scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -719,20 +702,18 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val noteList = reporter.noteProvidedEventsReceived
-      assert(noteList.size == 1)
-      assert(noteList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val noteList = reporter.noteProvidedEventsReceived
+        assert(noteList.size == 1)
+        assert(noteList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -745,21 +726,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
 
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in feature body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -776,21 +755,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
 
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -806,20 +783,17 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-
-      val alertList = reporter.alertProvidedEventsReceived
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send an AlertProvided event for an alert in Future returned by scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -837,20 +811,18 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val alertList = reporter.alertProvidedEventsReceived
-      assert(alertList.size == 1)
-      assert(alertList(0).message == "hi there")
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val alertList = reporter.alertProvidedEventsReceived
+        assert(alertList.size == 1)
+        assert(alertList(0).message == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -863,21 +835,19 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
 
-      assert(markupList.size == 1)
-      assert(markupList(0).text == "hi there")
+        assert(markupList.size == 1)
+        assert(markupList(0).text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in feature body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -894,21 +864,18 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
 
-      val markupList = reporter.markupProvidedEventsReceived
-
-      assert(markupList.size == 1)
-      assert(markupList(0).text == "hi there")
+        assert(markupList.size == 1)
+        assert(markupList(0).text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -924,27 +891,25 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
-      assert(markupList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
+        assert(markupList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[MarkupProvided])
-      val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
-      assert(markupProvided.text == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[MarkupProvided])
+        val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
+        assert(markupProvided.text == "hi there")
+      }
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by scenario body") {
       class MySuite extends featurespec.FixtureAsyncFeatureSpecLike  {
-
-        //SCALATESTJS-ONLY implicit override def executionContext = org.scalatest.concurrent.TestExecutionContext.runNow
-//SCALATESTNATIVE-ONLY implicit override def executionContext = scala.concurrent.ExecutionContext.Implicits.global
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -962,20 +927,21 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new MySuite
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
 
-      val markupList = reporter.markupProvidedEventsReceived
-      assert(markupList.size == 0)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { repo =>
+        val markupList = reporter.markupProvidedEventsReceived
+        assert(markupList.size == 0)
 
-      val testSucceededList = reporter.testSucceededEventsReceived
-      assert(testSucceededList.size == 1)
-      assert(testSucceededList(0).recordedEvents.size == 1)
-      val recordedEvent = testSucceededList(0).recordedEvents(0)
-      assert(recordedEvent.isInstanceOf[MarkupProvided])
-      val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
-      assert(markupProvided.text == "hi there")
+        val testSucceededList = reporter.testSucceededEventsReceived
+        assert(testSucceededList.size == 1)
+        assert(testSucceededList(0).recordedEvents.size == 1)
+        val recordedEvent = testSucceededList(0).recordedEvents(0)
+        assert(recordedEvent.isInstanceOf[MarkupProvided])
+        val markupProvided = recordedEvent.asInstanceOf[MarkupProvided]
+        assert(markupProvided.text == "hi there")
+      }
     }
 
     it("should allow other execution context to be used") {
@@ -1009,14 +975,14 @@ class AsyncFeatureSpecLikeSpec extends scalatest.funspec.AnyFunSpec {
       val suite = new TestSpec
       val reporter = new EventRecordingReporter
       val status = suite.run(None, Args(reporter))
-
-      // SKIP-SCALATESTJS,NATIVE-START
-      status.waitUntilCompleted()
-      // SKIP-SCALATESTJS,NATIVE-END
-      assert(reporter.scopeOpenedEventsReceived.length == 3)
-      assert(reporter.scopeClosedEventsReceived.length == 3)
-      assert(reporter.testStartingEventsReceived.length == 3)
-      assert(reporter.testSucceededEventsReceived.length == 3)
+      val promise = Promise[EventRecordingReporter]
+      status whenCompleted { _ => promise.success(reporter) }
+      promise.future.map { r =>
+        assert(reporter.scopeOpenedEventsReceived.length == 3)
+        assert(reporter.scopeClosedEventsReceived.length == 3)
+        assert(reporter.testStartingEventsReceived.length == 3)
+        assert(reporter.testSucceededEventsReceived.length == 3)
+      }
     }
 
   }
